Add a molnums topology attribute to TPR parser
PR #1557 introduced the moltype attribute that gives a name to each kind
of molecule in a topology. This commit adds the related molnums
attribute that gives an index to each individual molecules.

The molnums attributes allows to split a system by molecule without
ambiguity when the attribute is defines. Without this commit, the only
way of splitting a system per molecule is `atom_group.fragments`. While
it works most of the time, it can fail in some topology that use other
topology tricks than bonds to connect atoms; this is the case in gromacs
with virtual sites for instance. Fragments can also be time consuming on
large systems. This commits allows to slit a system per molecule without
ambiguity with `atom_group.groupby('molnums')` or
`atom_group.split('molecule')`.

This commit also adds the `molnum` selection keyword. To do so, this
commit refactored `mda.core.selection.SelectionRange` to avoid code
duplication.