enable pytest warnings with a permissive filter
This will print all warnings that pytest collects. Since we have some warnings
that we always issues as information for a user they don't give use as much
information as seeing that we call deprecated functions in our dependencies. To
silence the warnings log we filter those warnings out.

Now mostly deprecation warnings are left.